fix(static-pages): Pages are being generated statically
