+++ web/package.json
@@ -8,7 +8,8 @@
     "build": "next build",
     "lint": "eslint .",
     "lint:fix": "eslint . --fix",
-    "start": "next start"
+    "start": "next start",
+    "debug": "NODE_OPTIONS='--inspect' next dev"
   },
   "dependencies": {
     "@portabletext/react": "^1.0.2",
