+++ web/pages/[[...slug]].tsx
@@ -1,13 +1,59 @@
-import imageUrlBuilder from '@sanity/image-url';
-import groq from 'groq';
+/* eslint-disable @typescript-eslint/no-unused-vars */
+import { GetServerSideProps, GetStaticPaths, GetStaticProps } from 'next';
 import { NextSeo } from 'next-seo';
-import { GetServerSideProps } from 'next';
+import groq from 'groq';
+import imageUrlBuilder from '@sanity/image-url';
 
 import client from '../client';
 import Layout from '../components/Layout';
 import RenderSections from '../components/RenderSections';
 import { getSlugVariations, slugParamToPath } from '../utils/urls';
-import type { ConfigProps } from './_app';
+
+interface FooterNavigationProps {
+  _id: string;
+  title: string;
+  page: object;
+  slug: {
+    current: string;
+  };
+}
+
+export interface ConfigProps {
+  _id: string;
+  _rev: string;
+  _type: string;
+  _createdAt: string;
+  _updatedAt: string;
+  title: string;
+  mainNavigation: {
+    _id: string;
+    _rev: string;
+    _type: string;
+    title: string;
+    slug: {
+      current: string;
+    };
+  }[];
+  footerNavigation: FooterNavigationProps[];
+  footerText: {
+    _key: string;
+    _type: string;
+    title: string;
+    slug: string;
+    children: any[];
+    markDefs: any[];
+    style: string;
+  }[];
+  logo: {
+    logo: string;
+    title: string;
+    asset: {
+      url: string;
+      extension?: string;
+    };
+  };
+  url: string;
+}
 
 const pageFragment = groq`
 ...,
@@ -23,12 +69,27 @@ content[] {
   }
 }`;
 
+const siteConfigQuery = groq`
+  *[_id == "global-config"] {
+    ...,
+    logo {asset->{extension, url}},
+    mainNavigation[] -> {
+      ...,
+      "title": page->title
+    },
+    footerNavigation[] -> {
+      ...,
+      "title": page->title
+    }
+  }[0]
+  `;
+
 /**
  * Fetches data for our pages.
  *
  * The [[...slug]] name for this file is intentional - it means Next will run this getServerSideProps
  * for every page requested - /, /about, /contact, etc..
- * From the received params.slug, we're able to query Sanity for the route coresponding to the currently requested path.
+ * From the received params.slug, we're able to query Sanity for the route corresponding to the currently requested path.
  */
 
 interface ServerProps extends GetServerSideProps {
@@ -37,8 +98,26 @@ interface ServerProps extends GetServerSideProps {
   };
 }
 
-export const getServerSideProps = async ({ params }: ServerProps) => {
-  const slug = slugParamToPath(params?.slug);
+export const getStaticPaths: GetStaticPaths = async () => {
+  const { allRoutesSlugs } = await client.fetch(groq`{
+    // get all documents that are type of route as long as they are not a draft
+    "allRoutesSlugs": *[
+      _type == "route" &&
+      !(_id in path("drafts.**"))
+    ].slug.current,
+  }`);
+
+  const allRoutes = allRoutesSlugs.map((slug: string) => (slug === '/' ? '/' : `/${slug}`));
+
+  return {
+    paths: allRoutes,
+    fallback: true,
+  };
+};
+
+export const getStaticProps: GetStaticProps = async (props) => {
+  const slug = slugParamToPath(props.params?.slug);
+  const config = await client.fetch(siteConfigQuery);
 
   let data;
 
@@ -93,43 +172,35 @@ interface LandingPageProps {
   config: ConfigProps;
 }
 
-const LandingPage = ({
-  title = 'Missing title',
-  description,
-  disallowRobots,
-  openGraphImage,
-  content = [],
-  config,
-  slug,
-}: LandingPageProps) => {
+const LandingPage = ({ content, openGraphImage, ...rest }: LandingPageProps) => {
   const openGraphImages = openGraphImage
     ? [
         {
           url: builder.image(openGraphImage).width(800).height(600).url(),
           width: 800,
           height: 600,
-          alt: title,
+          alt: rest.title,
         },
         {
           // Facebook recommended size
           url: builder.image(openGraphImage).width(1200).height(630).url(),
           width: 1200,
           height: 630,
-          alt: title,
+          alt: rest.title,
         },
         {
           // Square 1:1
           url: builder.image(openGraphImage).width(600).height(600).url(),
           width: 600,
           height: 600,
-          alt: title,
+          alt: rest.title,
         },
       ]
     : [];
 
   return (
-    <Layout config={config}>
-      <NextSeo
+    <>
+      {/* <NextSeo
         title={title}
         titleTemplate={`%s | ${config.title}`}
         description={description}
@@ -138,9 +209,9 @@ const LandingPage = ({
           images: openGraphImages,
         }}
         noindex={disallowRobots}
-      />
+      /> */}
       {content && <RenderSections sections={content} />}
-    </Layout>
+    </>
   );
 };
